add  return表 分页 ID 查询 修改 和前端  returnhistory 表的 ID查询

+++ jcart-administration-back/src/main/java/io/wzh/jcartadministrationback/dto/in/OrderHistoryCreateInDTO.java
@@ -6,7 +6,7 @@ public class OrderHistoryCreateInDTO {
     private String comment;
     private Boolean customerNotified;
 
-    public Integer getOrderId() {
+    public Long getOrderId() {
         return orderId;
     }
